requirements.txt.txt


# --- requirements for training data capture ---
windows-capture
pillow
pynput

# --- ---